% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abadie_v.R
\name{abadie_v}
\alias{abadie_v}
\title{Abadie's nu}
\usage{
abadie_v(z, a, ps)
}
\arguments{
\item{z}{numeric A binary instrument, e.g., randomized treatment assignment.}

\item{a}{numeric The actual received treatment, must be the same length as \code{z}.}

\item{ps}{numeric The propensity scores \eqn{\Pr(Z = 1 \mid X)}.}
}
\value{
numeric The formula defined above.
}
\description{
This function computes the derivative of Abadie's kappa,
\deqn{\nu(\pi)=\dot{\kappa}(\pi)=\frac{Z(1-A)}{\pi^{2}}-\frac{A(1-Z)}{(1-\pi)^{2}}}.
}
\examples{
abadie_v(1, 1, 0.5)
abadie_v(1, 0, 0.5)
abadie_v(0, 1, 0.5)
abadie_v(0, 0, 0.5)
abadie_v(c(1, 1, 0, 0), c(1, 0, 1, 0), 0.5)
abadie_v(c(1, 1), c(1, 1), c(0.4, 0.6))
abadie_v(c(1, 1), c(0, 0), c(0.4, 0.6))
}
