% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_internal.R
\name{dgp_internal}
\alias{dgp_internal}
\title{Data generating process: linear transformation model with non-compliance}
\usage{
dgp_internal(
  n = 200,
  x1_dist = "normal",
  x2_dist = "normal",
  sigma_x = 1,
  nonlinearity = "sin",
  ps_link = "logit",
  ps_trim = 0.05,
  eta_0 = 0,
  eta_1 = 0,
  eta_2 = 0,
  eta_3 = 0,
  eta_4 = 0,
  p_c = 0.7,
  rho_a = 0.5,
  nc_scale = 0.7,
  nc_shift = 1,
  alpha = 1,
  beta_1 = 0.5,
  beta_2 = -0.7,
  beta_3 = 0,
  sigma_e = 1,
  error_dist = "gumbel"
)
}
\arguments{
\item{n}{numeric Number of samples (observations).}

\item{x1_dist}{character {"binary", "uniform", "normal"}.}

\item{x2_dist}{character {"binary", "uniform", "normal"}.}

\item{sigma_x}{numeric Standard deviation of x (if not binary).}

\item{nonlinearity}{character Apply what nonlinear transformation to \code{x}?}

\item{ps_link}{character link function for propensity score DGP, see ?make.link.}

\item{ps_trim}{numeric This enforces positivity assumption.
(ps_trim, 1 - ps_trim) will be the bound of generated
propensity score, implemented by rejection sampling.}

\item{eta_0}{numeric Coefficients for the propensity score DGP.}

\item{eta_1}{numeric Coefficients for the propensity score DGP.}

\item{eta_2}{numeric Coefficients for the propensity score DGP.}

\item{eta_3}{numeric Coefficients for the propensity score DGP.}

\item{eta_4}{numeric Coefficients for the propensity score DGP.}

\item{p_c}{numeric Compliance rate.}

\item{rho_a}{numeric Proportion of always-takers among non-compliers.}

\item{nc_scale}{numeric Scale non-complier coefficients.}

\item{nc_shift}{numeric Shift non-complier outcomes.}

\item{alpha}{numeric Coefficient associated with the treatment difference}

\item{beta_1}{numeric Coefficients associated with \code{x1}.}

\item{beta_2}{numeric Coefficients associated with \code{x2}.}

\item{beta_3}{numeric Coefficients associated with \code{x1 * x2}.}

\item{sigma_e}{standard deviation of the error term, default is 1.}

\item{error_dist}{"gumbel" or "normal".}
}
\value{
df - generated data frame with columns {x1, x2, x2_t, lp, ps, z, c, a, ey, y} \cr
df$... - see "dgp_prop_score()" \cr
df$z - randomized treatment assignment based on the propensity scores \cr
df$c - complier status {0, 1, 2} = {always-taker, complier, never-taker} \cr
df$a - actual treatment received \cr
df$ey - expected value of y \cr
df$y - outcome \cr
}
\description{
TODO: ADD MORE DESCRIPTION HERE.
}
