% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_obs.R
\name{dgp_obs}
\alias{dgp_obs}
\title{Data generating process: observational study}
\usage{
dgp_obs(
  n = 200,
  p_c = 0.7,
  alpha = 1,
  beta_1 = 0.5,
  beta_2 = -0.7,
  beta_3 = 0,
  error_dist = "gumbel"
)
}
\arguments{
\item{n}{numeric Number of samples (observations).}

\item{p_c}{numeric Compliance rate.}

\item{alpha}{numeric Coefficient associated with the treatment difference}

\item{beta_1}{numeric Coefficients associated with \code{x1}.}

\item{beta_2}{numeric Coefficients associated with \code{x2}.}

\item{beta_3}{numeric Coefficients associated with \code{x1 * x2}.}

\item{error_dist}{"gumbel" or "normal".}
}
\value{
TODO
}
\description{
This function inherit dgp_internal(). TODO: ADD MORE DESCRIPTION.
}
\examples{
dgp_obs(n = 6)
}
