% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp_prop_score.R
\name{dgp_prop_score}
\alias{dgp_prop_score}
\title{Data generating process: propensity score}
\usage{
dgp_prop_score(
  n = 200,
  x1_dist = "normal",
  x2_dist = "normal",
  sigma_x = 1,
  nonlinearity = "sin",
  ps_link = "logit",
  ps_trim = 0.05,
  eta_0 = 0,
  eta_1 = 0,
  eta_2 = 0,
  eta_3 = 0,
  eta_4 = 0
)
}
\arguments{
\item{n}{numeric Number of samples (observations).}

\item{x1_dist}{character {"binary", "uniform", "normal"}.}

\item{x2_dist}{character {"binary", "uniform", "normal"}.}

\item{sigma_x}{numeric Standard deviation of x (if not binary).}

\item{nonlinearity}{character Apply what nonlinear transformation to \code{x}?}

\item{ps_link}{character link function for propensity score DGP, see ?make.link.}

\item{ps_trim}{numeric This enforces positivity assumption.
(ps_trim, 1 - ps_trim) will be the bound of generated
propensity score, implemented by rejection sampling.}

\item{eta_0}{numeric Coefficients for the propensity score DGP.}

\item{eta_1}{numeric Coefficients for the propensity score DGP.}

\item{eta_2}{numeric Coefficients for the propensity score DGP.}

\item{eta_3}{numeric Coefficients for the propensity score DGP.}

\item{eta_4}{numeric Coefficients for the propensity score DGP.}
}
\value{
data.frame \cr
\code{df} - generated data frame with columns {x1, x2, x2_t, lp, ps, z} \cr
\code{df$x1} - "binary" ~ Bernoulli(0.5), "uniform" ~ U(-sqrt(3)sigma_x, sqrt(3)sigma_x), "normal" ~ N(0, sigma_x^2) \cr
\code{df$x2} - "uniform" ~ U(-sqrt(3)sigma_x, sqrt(3)sigma_x), "normal" ~ N(0, sigma_x^2) \cr
\code{df$x1_t} - "sin": sin(x1), "inv": sign(x1) / (abs(x1) + 1) \cr
\code{df$x2_t} - "sin": sin(x2), "inv": sign(x2) / (abs(x2) + 1) \cr
\code{df$lp} - linear part: \code{eta_0 + eta_1 * x1 + eta_2 * x2 + eta_3 * x1_t + eta_4 * x2_t} \cr
\code{df$ps} = \code{stats::make.link(ps_link)$linkinv(eta_0 + eta_1 * x1 + eta_2 * x2 + eta_3 * x1_t + eta_4 * x2_t)} \cr
\code{df$z} - randomized treatment assignment based on the propensity scores.
}
\description{
Generates the covariates and the propensity scores. The propensity scores are
generated by
\code{ps = stats::make.link(ps_link)$linkinv(eta_0 + eta_1 * x1 + eta_2 * x2 + eta_3 * x1_t + eta_4 * x2_t)}.
}
\examples{
set.seed(42)
dgp_prop_score(n = 6)
}
