% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate Data for Simulation Studies}
\usage{
generate_data(
  n_obs = 200,
  ps_type = c("la", "nla", "nlna", "omit", "complex", "rct"),
  p_c = 0.7,
  alpha = 1,
  beta_1 = 0.5,
  beta_2 = -0.7,
  beta_3 = 0,
  error_dist = c("gumbel", "normal")
)
}
\arguments{
\item{n_obs}{numeric Number of observations.}

\item{ps_type}{character Type of IV propensity score DGP.}

\item{p_c}{numeric Compliance rate.}

\item{alpha}{numeric Coefficient associated with the treatment difference.}

\item{beta_1}{numeric Coefficients associated with \code{x1}.}

\item{beta_2}{numeric Coefficients associated with \code{x2}.}

\item{beta_3}{numeric Coefficients associated with \code{x1 * x2}.}

\item{error_dist}{"gumbel" or "normal".}
}
\value{
A data frame
}
\description{
Generate Data for Simulation Studies
}
\examples{
set.seed(42)
generate_data(n_obs = 6, ps_type = "la")
}
