% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_IF.R
\name{glm_IF}
\alias{glm_IF}
\title{Empirical influence function of glm object}
\usage{
glm_IF(glm_obj)
}
\arguments{
\item{glm_obj}{The object returned by glm().}
}
\value{
matrix A \code{n} by \code{p} matrix.
}
\description{
Empirical influence function of glm object
}
\examples{
x1 <- rnorm(1e3)
x2 <- rnorm(1e3)
y <- rbinom(n = 1e3, size = 1, plogis(-1 + x1 + 2 * x2))
glm_obj <- glm(y ~ x1 + x2, family = binomial(link = "logit"))
all.equal(crossprod(glm_IF(glm_obj)) / 1e6, sandwich::vcovHC(glm_obj, type = "HC"))
}
