% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivpim.R
\name{ivpim}
\alias{ivpim}
\title{Instrumental variable probabilistic index model}
\usage{
ivpim(
  y,
  z,
  a,
  X = NULL,
  ps_model,
  link = "logit",
  init = NULL,
  tol = sqrt(.Machine$double.eps),
  max.iter = 100,
  nleqslv.global = "none",
  trace = FALSE,
  test.nleqslv = FALSE,
  keep.data = FALSE
)
}
\arguments{
\item{y}{numeric Outcome vector.}

\item{z}{numeric Binary instrument, e.g., randomized treatment assignment.}

\item{a}{numeric Actual received treatment, must be the same length as \code{z}.}

\item{X}{numeric Covariates.}

\item{ps_model}{glm First step propensity score \code{glm()} object.}

\item{link}{character Link function: "logit", or "probit".}

\item{init}{numeric Initial guess of Newton's method.}

\item{tol}{numeric Numeric tolerance of \code{nleqslv()}.}

\item{max.iter}{numeric Maximum iteration of Newton's method.}

\item{nleqslv.global}{character The global strategy for Newton's method. See ?nleqslv::nleqslv.}

\item{trace}{logical Show Newton's method iteration report if TRUE.}

\item{test.nleqslv}{logical Test different global strategies for Newton's method if TRUE. See ?nleqslv::testnslv.}

\item{keep.data}{logical Should the returned object keep original data?}
}
\value{
A list containing the estimated coefficients and their covaraince
matrix. It also contains the diagnostics of \code{nleqlsv()} procedure.
If \code{keep.data} is \code{TRUE}, then the inputs \code{y}, \code{z}, \code{a}, \code{X}, \code{link}, \code{w}
will also be returned.
}
\description{
This model targets the compatible complier probabilistic index model.
}
