% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance_test.R
\name{new_compliance_test}
\alias{new_compliance_test}
\title{Create a compliance test}
\usage{
new_compliance_test(desc, code, ref_fn, opts = list())
}
\arguments{
\item{desc}{Compliance test description.}

\item{code}{Compliance test code.}

\item{ref_fn}{A reference implementation, i.e. a function that passes
the test. This is used as default argument value for the returned function
(see Value).}

\item{opts}{Optional arguments for tuning the test behavior.}
}
\value{
A \link{function} that runs the test. The first argument is a function
candidate at complying with the test. Following arguments are those
passed in \code{opts}.
}
\description{
A compliance test is a unit test whose expectations are defined around
a compliant function. Use the pronoun \code{.f} in test \code{code} to refer
to this function.
}
